+++ src/main/java/io/github/astrapi69/design/pattern/state/wizard/model/WizardStateInfo.java
@@ -1,0 +1,57 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2024 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.design.pattern.state.wizard.model;
+
+import lombok.AccessLevel;
+import lombok.Data;
+import lombok.experimental.FieldDefaults;
+import lombok.experimental.SuperBuilder;
+
+/**
+ * The class {@link WizardStateInfo} represents the state information of a wizard step. It holds
+ * information about the step's name and whether it is the first, last, has a previous step, or has
+ * a next step.
+ */
+@Data
+@SuperBuilder(toBuilder = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class WizardStateInfo
+{
+
+	/** The name of the wizard step */
+	String name;
+
+	/** Flag indicating if this is the first step */
+	boolean first;
+
+	/** Flag indicating if this is the last step */
+	boolean last;
+
+	/** Flag indicating if there is a previous step */
+	boolean previous;
+
+	/** Flag indicating if there is a next step */
+	boolean next;
+}