+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 Version 7.1-SNAPSHOT
 -------------
 
+ADDED:
+
+- new data info class WizardStateInfo with all information from the wizard state object
+- new methods in the interface WizardState for interact with the WizardStateInfo object
+
+CHANGED:
+
+- update gradle to new version 8.10-rc-1
+- update gradle-plugin with the id io.freefair.lombok to new version 8.7.1
 
 Version 7
 -------------